backend/requirements.txt
Flask
Flask-Cors
google-generativeai
google-auth
google-api-python-client
gunicorn
python-dotenv